version 0.0.1

+++ lib.nix
@@ -1,7 +1,19 @@
-{ pkgs, ... }:
-{
+{ pkgs, lib, ... }:
+rec {
   modules = {
     hostKeys = ./modules/hostKeys;
+  };
+
+  types = {
+    key = pkgs.lib.mkOptionType {
+      name = "key";
+      check = (v: builtins.isAttrs v && builtins.hasAttr "type" v && builtins.elem v.type ["literal" "path"]);
+    };
+  };
+
+  literal = val: {
+    type = "literal";
+    value = val;
   };
 
   crypts = {
@@ -39,16 +51,23 @@
 
   inputs = {
     single =
-      { var, prompt, ... }:
+      { var, prompt }:
       ''
-        echo -n "${prompt}"
+        echo -n "${prompt}: "
         read -r ${var}
       '';
-    mutli =
-      { var, prompt, ... }:
+    multi =
+      { var, prompt }:
       ''
-        echo -n "${prompt}"
-        $${var}=$(</dev/stdin)
+        echo -n "${prompt}: "
+        ${var}=$(</dev/stdin)
+      '';
+    
+    yesNo =
+      { var, prompt }:
+      ''
+        ${inputs.single { inherit var; prompt = "${prompt} [y/N]"; }}
+        ${var}="$(echo "''$${var}" | tr '[:upper:]' '[:lower:]')"
       '';
   };
 
@@ -58,12 +77,26 @@
         publicFile ? "public",
         privateFile ? "private",
         type ? "ed25519",
+        promptExisting ? false,
         ...
       }:
       ''
-        ${pkgs.openssh}/bin/ssh-keygen -t ${type} -N "" -C "" -q -f key
-        mv key.pub "$out/${publicFile}"
-        mv key "$out/${privateFile}"
+        existing=n
+        ${
+          lib.optionalString promptExisting
+          (inputs.yesNo { var = "existing"; prompt = "Use existing key?"; })
+        }
+
+        if [ "$existing" == "y" ]; then
+          ${inputs.single { var = "public"; prompt = "Public key"; }}
+          ${inputs.multi { var = "private"; prompt = "Private key"; }}
+          echo "$public" > "$out/${publicFile}"
+          echo "$private" > "$out/${privateFile}"
+        else
+          ${pkgs.openssh}/bin/ssh-keygen -t ${type} -N "" -C "" -q -f key
+          mv key.pub "$out/${publicFile}"
+          mv key "$out/${privateFile}"
+        fi
       '';
   };
 }
